Converted remaining screens to new organization, more cleanup WIP
