+++ app/src/main/java/com/brandoncano/resistancecalculator/data/NavigationBarItemPO.kt
@@ -1,0 +1,8 @@
+package com.brandoncano.resistancecalculator.data
+
+import androidx.compose.ui.graphics.vector.ImageVector
+
+/**
+ * Job: Holds the content for bottom navigation bar option
+ */
+data class NavigationBarItemPO(val label: String, val imageVector: ImageVector)
