+++ app/src/main/java/com/brandoncano/resistancecalculator/navigation/calculators/ColorToValueNavEntry.kt
@@ -18,7 +18,7 @@ import com.brandoncano.resistancecalculator.model.ResistorCtvViewModel
 import com.brandoncano.resistancecalculator.navigation.Screen
 import com.brandoncano.resistancecalculator.navigation.navigateToAbout
 import com.brandoncano.resistancecalculator.navigation.navigateToColorCodeIec
-import com.brandoncano.resistancecalculator.ui.screens.ctv.ColorToValueScreen
+import com.brandoncano.resistancecalculator.ui.screens.calculators.ColorToValueScreen
 import com.brandoncano.resistancecalculator.util.share.ShareResistor
 import com.brandoncano.sharedcomponents.utils.SendFeedback
 import com.brandoncano.sharedcomponents.utils.ShareText
