+++ app/src/main/java/com/brandoncano/resistancecalculator/navigation/calculators/ParallelCalculatorNavEntry.kt
@@ -1,4 +1,4 @@
-package com.brandoncano.resistancecalculator.navigation.circuit
+package com.brandoncano.resistancecalculator.navigation.calculators
 
 import androidx.compose.animation.EnterTransition
 import androidx.compose.animation.ExitTransition
@@ -19,14 +19,14 @@ import com.brandoncano.resistancecalculator.model.CircuitViewModel
 import com.brandoncano.resistancecalculator.navigation.Screen
 import com.brandoncano.resistancecalculator.navigation.navigateToAbout
 import com.brandoncano.resistancecalculator.navigation.navigateToCircuitEquations
-import com.brandoncano.resistancecalculator.ui.screens.circuit.CircuitCalculatorScreen
+import com.brandoncano.resistancecalculator.ui.screens.calculators.CircuitCalculatorScreen
 import com.brandoncano.sharedcomponents.utils.SendFeedback
 
-fun NavGraphBuilder.seriesCalculatorScreen(
+fun NavGraphBuilder.parallelCalculatorScreen(
     navHostController: NavHostController,
 ) {
     composable(
-        route = Screen.SeriesCalculator.route,
+        route = Screen.ParallelCalculator.route,
         enterTransition = { slideInVertically(initialOffsetY = { it }) },
         exitTransition = { ExitTransition.None },
         popEnterTransition = { EnterTransition.None },
@@ -34,15 +34,15 @@ fun NavGraphBuilder.seriesCalculatorScreen(
     ) {
         val context = LocalContext.current
         val focusManager = LocalFocusManager.current
-        val viewModel: CircuitViewModel = viewModel()
+        val viewModel: CircuitViewModel = viewModel<CircuitViewModel>()
         val circuit by viewModel.circuitStateTOStateFlow.collectAsState()
         // We need to call this update because of shared preferences being used for both
-        viewModel.updateValues(circuit.isSameValues, circuit.resistorCount, circuit.units, true)
+        viewModel.updateValues(circuit.isSameValues, circuit.resistorCount, circuit.units, false)
 
         CircuitCalculatorScreen(
-            circuitTitle = R.string.circuit_title_series,
-            circuitVector = R.drawable.img_resistors_series,
-            vectorSize = Pair(390.dp, 106.dp),
+            circuitTitle = R.string.circuit_title_parallel,
+            circuitVector = R.drawable.img_resistors_parallel,
+            vectorSize = Pair(301.dp, 177.dp),
             circuit = circuit,
             onNavigateBack = { navHostController.popBackStack() },
             onClearSelectionsTapped = {
@@ -51,8 +51,8 @@ fun NavGraphBuilder.seriesCalculatorScreen(
             },
             onFeedbackTapped = { SendFeedback.execute(context, Links.APP_NAME) },
             onAboutTapped = { navigateToAbout(navHostController) },
-            onValueChanged = { sameValues, resistorCount, units ->
-                viewModel.updateValues(sameValues, resistorCount, units, true)
+            onOptionSelected = { sameValues, resistorCount, units ->
+                viewModel.updateValues(sameValues, resistorCount, units, false)
             },
             onLearnCircuitEquationsTapped = { navigateToCircuitEquations(navHostController) },
         )
