+++ app/src/main/java/com/brandoncano/resistancecalculator/ui/screens/calculators/SmdScreen.kt
@@ -1,16 +1,16 @@
-package com.brandoncano.resistancecalculator.ui.screens.smd
+package com.brandoncano.resistancecalculator.ui.screens.calculators
 
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.safeDrawing
-import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.outlined.Explicit
@@ -18,8 +18,8 @@ import androidx.compose.material.icons.outlined.Looks3
 import androidx.compose.material.icons.outlined.Looks4
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -27,15 +27,17 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.res.dimensionResource
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardCapitalization
 import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import com.brandoncano.resistancecalculator.R
 import com.brandoncano.resistancecalculator.constants.Lists
+import com.brandoncano.resistancecalculator.data.NavigationBarItemPO
 import com.brandoncano.resistancecalculator.to.SmdResistor
 import com.brandoncano.resistancecalculator.ui.composables.AboutAppMenuItem
 import com.brandoncano.resistancecalculator.ui.composables.AppTextField
@@ -47,12 +49,18 @@ import com.brandoncano.resistancecalculator.ui.composables.MenuIconButton
 import com.brandoncano.resistancecalculator.ui.composables.ShareImageMenuItem
 import com.brandoncano.resistancecalculator.ui.composables.ShareTextMenuItem
 import com.brandoncano.resistancecalculator.ui.composables.TextDropDownMenu
+import com.brandoncano.resistancecalculator.ui.composables.m3.M3DisplayCard
+import com.brandoncano.resistancecalculator.ui.composables.m3.M3NavigationBar
 import com.brandoncano.resistancecalculator.ui.composables.m3.M3OutlinedCard
+import com.brandoncano.resistancecalculator.ui.composables.m3.M3Scaffold
+import com.brandoncano.resistancecalculator.ui.composables.m3.M3ScreenColumn
 import com.brandoncano.resistancecalculator.ui.composables.m3.M3TopAppBar
 import com.brandoncano.resistancecalculator.ui.theme.ResistorCalculatorTheme
-import com.brandoncano.sharedcomponents.composables.AppNavigationBar
+import com.brandoncano.resistancecalculator.ui.theme.black
+import com.brandoncano.resistancecalculator.ui.theme.resistor_wire
+import com.brandoncano.resistancecalculator.ui.theme.white
+import com.brandoncano.resistancecalculator.util.resistor.formatResistance
 import com.brandoncano.sharedcomponents.composables.AppScreenPreviews
-import com.brandoncano.sharedcomponents.data.NavigationBarOptions
 import java.util.Locale
 
 @OptIn(ExperimentalMaterial3Api::class) // For TopAppBar
@@ -72,10 +80,7 @@ fun SmdScreen(
     onLearnSmdCodesTapped: () -> Unit,
 ) {
     var showMenu by remember { mutableStateOf(false) }
-    val scrollBehavior = TopAppBarDefaults.pinnedScrollBehavior()
-
-    Scaffold(
-        modifier = Modifier.nestedScroll(scrollBehavior.nestedScrollConnection),
+    M3Scaffold(
         topBar = {
             M3TopAppBar(
                 titleText = stringResource(R.string.smd_title),
@@ -111,30 +116,29 @@ fun SmdScreen(
                         }
                     }
                 },
-                scrollBehavior = scrollBehavior,
+                scrollBehavior = it,
             )
         },
         bottomBar = {
-            AppNavigationBar(
+            M3NavigationBar(
                 selection = resistor.navBarSelection,
                 onClick = { onNavBarSelectionChanged(it) },
                 options = listOf(
-                    NavigationBarOptions(
+                    NavigationBarItemPO(
                         label = stringResource(id = R.string.smd_navbar_three_eia),
                         imageVector = Icons.Outlined.Looks3,
                     ),
-                    NavigationBarOptions(
+                    NavigationBarItemPO(
                         label = stringResource(id = R.string.smd_navbar_four_eia),
                         imageVector = Icons.Outlined.Looks4,
                     ),
-                    NavigationBarOptions(
+                    NavigationBarItemPO(
                         label = stringResource(id = R.string.smd_navbar_eia_96),
                         imageVector = Icons.Outlined.Explicit,
                     ),
                 ),
             )
         },
-        contentWindowInsets = WindowInsets.safeDrawing,
     ) { paddingValues ->
         SmdScreenContent(
             paddingValues = paddingValues,
@@ -156,13 +160,8 @@ private fun SmdScreenContent(
     onOptionSelected: (String) -> Unit,
     onLearnSmdCodesTapped: () -> Unit,
 ) {
-    val sidePadding = dimensionResource(R.dimen.app_side_padding)
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .verticalScroll(rememberScrollState())
-            .padding(paddingValues)
-            .padding(horizontal = sidePadding),
+    M3ScreenColumn(
+        paddingValues = paddingValues,
         horizontalAlignment = Alignment.CenterHorizontally,
     ) {
         Spacer(modifier = Modifier.padding(top = 32.dp))
@@ -200,6 +199,60 @@ private fun SmdScreenContent(
         }
         BottomScreenSpacer()
     }
+}
+
+@Composable
+private fun SmdResistorLayout(resistor: SmdResistor, isError: Boolean, verticalPadding: Dp = 0.dp) {
+    val code = if (isError) {
+        stringResource(id = R.string.error_na)
+    }  else {
+        resistor.code
+    }
+    val resistance = when {
+        resistor.isEmpty() -> stringResource(id = R.string.smd_default_value)
+        isError -> stringResource(id = R.string.error_na)
+        else -> resistor.formatResistance()
+    }
+    Column(
+        modifier = Modifier.padding(horizontal = 32.dp, vertical = verticalPadding),
+        horizontalAlignment = Alignment.CenterHorizontally,
+    ) {
+        SmdResistorImage {
+            Text(
+                text = code,
+                style = MaterialTheme.typography.headlineLarge.white(),
+            )
+        }
+        Spacer(modifier = Modifier.height(16.dp))
+        M3DisplayCard(text = resistance)
+    }
+}
+
+@Composable
+private fun SmdResistorImage(content: @Composable (BoxScope.() -> Unit)) {
+    Box(
+        modifier = Modifier
+            .width(pxToDp(512f))
+            .height(pxToDp(224f))
+            .clip(RoundedCornerShape(4.dp))
+            .background(color = resistor_wire),
+        contentAlignment = Alignment.Center,
+    ) {
+        Box(
+            modifier = Modifier
+                .width(pxToDp(448f))
+                .height(pxToDp(200f))
+                .background(color = black),
+            contentAlignment = Alignment.Center,
+            content = content,
+        )
+    }
+}
+
+@Composable
+private fun pxToDp(px: Float): Dp {
+    val density = LocalDensity.current
+    return with(density) { px.toDp() }
 }
 
 @AppScreenPreviews
